+++ src/extension.ts
@@ -1,9 +1,9 @@
-import * as lh from './language_handler';
+import * as mbTaskExt from './language_handler';
 
 import * as vscode from 'vscode';
 
 export function activate(context: vscode.ExtensionContext) {
-	lh.initExtension(context);
+	mbTaskExt.initExtension(context);
 }
 
 export function deactivate() {}