+++ CHANGELOG.md
@@ -78,3 +78,13 @@ Encounter typescipt bug, VSC 1.94.2, Node.js 20.16.0, Npm 10.8.2
 ## 0.0.10
 
 - Terminal shell/icon
+
+## 0.11.1
+
+- Language definition save/load fix
+
+- Swift
+
+## 0.11.4
+
+- Language definition save/load fix, monitoring change to reload